Validaciones lógicas de recursos

+++ app/Http/Controllers/RawMaterialController.php
@@ -41,7 +41,7 @@ class RawMaterialController extends Controller
     public function store(Request $request)
     {
         if ( ! Auth::user()->can('rawMaterial_store')){
-            return redirect()->back()->withErrors(['warning' => 'No posees los permisos necesarios. Ponte en contacto con tu manager!.']);
+            return redirect()->back()->withErrors(['warning' => '¡No posees los permisos necesarios. Ponte en contacto con tu manager!.']);
         }
 
         $validando = $request->validate([
@@ -69,7 +69,7 @@ class RawMaterialController extends Controller
     public function update(Request $request)
     {
         if ( ! Auth::user()->can('rawMaterial_update')){
-            return redirect()->back()->withErrors(['warning' => 'No posees los permisos necesarios. Ponte en contacto con tu manager!.']);
+            return redirect()->back()->withErrors(['warning' => '¡No posees los permisos necesarios. Ponte en contacto con tu manager!.']);
         }
 
         if (!$request->rawMaterial_id) {
@@ -106,10 +106,13 @@ class RawMaterialController extends Controller
     public function destroy(Request $request)
     {
         if ( ! Auth::user()->can('rawMaterial_destroy')){
-            return redirect()->back()->withErrors(['warning' => 'No posees los permisos necesarios. Ponte en contacto con tu manager!.']);
+            return redirect()->back()->withErrors(['warning' => '¡No posees los permisos necesarios. Ponte en contacto con tu manager!.']);
         }
 
         $rawMaterial = RawMaterial::find($request->id);
+        if (count($rawMaterial->stoks)) {
+            return redirect()->back()->withErrors(['warning' => '¡No se puede eliminar este registro, dado que ya se encuentra relacionado con otras entradas en el sistema!.']);
+        }
         $rawMaterial->delete();
         return redirect()->back()->with('success', 'Registro eliminado correctamente!.');
     }
